+++ cmd/plugin.go
@@ -1,5 +1,5 @@
 package main
 
 import (
-	_ "github.com/mhchlib/mconfig/plugin/etcd"
+	_ "github.com/mhchlib/mconfig/pkg/plugin/etcd"
 )
