--- pkg/plugin.go
@@ -1,9 +1,0 @@
-package pkg
-
-func RegisterAppConfigStore(store AppConfigStore) error {
-	if appConfigStore != nil {
-		return Error_RepeateRegisterStore
-	}
-	appConfigStore = store
-	return nil
-}
